Project_2_(Grade_Calculator)
A program that illustrates a grade calculator that calculates and displays statistical information such as grades, hardest course among students and the easiest category for a student to achieve better grades during the course. In addition, the program makes sure with parameters every course selection to be precise and avoid user-input related issues. Lastly the program uses functions, lists and dictionaries in Python.